refactor: move entities and domain
