+++ domain/common.go
