+++ pkg/midtrans/midtrans_repository.go
@@ -1,7 +1,7 @@
 package midtrans
 
 import (
-	"Go-Starter-Template/pkg/entities"
+	"Go-Starter-Template/entities"
 	"context"
 	"gorm.io/gorm"
 )
